requirements.txt
torch>=2.0.0
numpy>=1.21.0
psycopg2-binary>=2.9.0
